scripts/content-pipeline/requirements.txt
python-dotenv
youtube-transcript-api
yt-dlp
trafilatura
anthropic
instructor
pydantic
typer
rich
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pyyaml
faster-whisper
feedparser
python-slugify
